Create Data-Driven_HomeRemodel_Insights_Analysis
The goal of this analysis is to provide insight for homeowners in Ames, Iowa looking to remodel a home, focusing on what might improve the overall appeal and value of certain features of the house to a potential buyer. House sale data was analyzed from 1,460 homes sold between 2006-2010 in Ames, Iowa (Kaggle, 2016), to identify the features (interior vs structural features) that house sale predictions appear to be most sensitive to.